+++ root/install_f1/schemas/_mssql_schema.sql
@@ -32,7 +32,7 @@ CREATE TABLE [phpbb_formel_drivers] (
 	[driver_name] [varchar] (255) DEFAULT ('') NOT NULL ,
 	[driver_img] [varchar] (255) DEFAULT ('') NOT NULL ,
 	[driver_team] [int] DEFAULT (0) NOT NULL ,
-	[driver_penalty] [int] DEFAULT (0) NOT NULL 
+	[driver_penalty] [float] DEFAULT (0) NOT NULL 
 ) ON [PRIMARY]
 GO
 
@@ -52,7 +52,7 @@ CREATE TABLE [phpbb_formel_teams] (
 	[team_name] [varchar] (255) DEFAULT ('') NOT NULL ,
 	[team_img] [varchar] (255) DEFAULT ('') NOT NULL ,
 	[team_car] [varchar] (255) DEFAULT ('') NOT NULL ,
-	[team_penalty] [int] DEFAULT (0) NOT NULL 
+	[team_penalty] [float] DEFAULT (0) NOT NULL 
 ) ON [PRIMARY]
 GO
